edit: edit update mechanism in blog admin

+++ app/Repositories/BlogRepository.php
@@ -69,18 +69,23 @@ class BlogRepository implements BlogInterface
 
     public function update($id, $data)
     {
+        // dd($data);
         $blog = $this->getById($id);
 
-        // delete old file
-        Storage::delete('public/blogs/thumbnail/' . $blog->thumbnail);
-        Storage::delete('public/blogs/main-image/' . $blog->main_image);
+        if (isset($data['thumbnail'])) {
+            $fileNameThumbnail = uniqid() . '.' . $data['thumbnail']->extension();
+            $data['thumbnail']->storeAs('public/blogs/thumbnail', $fileNameThumbnail);
+            Storage::delete('public/blogs/thumbnail/' . $blog->thumbnail);
+            $data['thumbnail'] = $fileNameThumbnail;
+        }
 
-        // upload new file
-        $fileNameThumbnail = uniqid() . '.' . $data['thumbnail']->extension();
-        $data['thumbnail']->storeAs('public/blogs/thumbnail', $fileNameThumbnail);
 
-        $fileNameMainImage = uniqid() . '.' . $data['main_image']->extension();
-        $data['main_image']->storeAs('public/blogs/main-image', $fileNameMainImage);
+        if (isset($data['main_image'])) {
+            $fileNameMainImage = uniqid() . '.' . $data['main_image']->extension();
+            $data['main_image']->storeAs('public/blogs/main-image', $fileNameMainImage);
+            Storage::delete('public/blogs/main-image/' . $blog->main_image);
+            $data['main_image'] = $fileNameMainImage;
+        }
 
         $slug = str_replace(' ', '-', strtolower($data['title']));
 
@@ -90,21 +95,21 @@ class BlogRepository implements BlogInterface
             $slug = $slug . '-' . uniqid();
         }
 
+        $published_date   = date('Y-m-d H:i:s');
+        $tag              = str_replace(',', ' - ', $data['tag']);
+        $meta_description = substr(strip_tags($data['content']), 0, 160);
+        $meta_keyword     = str_replace(',', '-', $data['tag']);
+
+
         DB::beginTransaction();
         try {
-            $blog->update([
-                'blog_category_id' => $data['blog_category_id'],
-                'thumbnail'        => $fileNameThumbnail,
-                'main_image'       => $fileNameMainImage,
-                'title'            => $data['title'],
-                'author_name'      => $data['author_name'],
-                'content'          => $data['content'],
-                'published_date'   => date('Y-m-d H:i:s'),
-                'tag'              => str_replace(',', ' - ', $data['tag']),
-                'meta_description' => substr(strip_tags($data['content']), 0, 160),
-                'meta_keyword'     => str_replace(',', '-', $data['tag']),
+            $blog->update(array_merge($data, [
+                'published_date'   => $published_date,
+                'tag'              => $tag,
+                'meta_description' => $meta_description,
+                'meta_keyword'     => $meta_keyword,
                 'slug'             => $slug
-            ]);
+            ]));
         } catch (\Throwable $th) {
             throw $th;
